Summary:
Commit https://github.com/facebook/react-native/commit/3d3b067f6fc831b6b23726087fe39cf39ef86f00 implemented zIndex support for Android. However after zIndex resorting of the views, the JS doesn't know about the new order. When dynamically removing the view from an array of views, JS sends the command to remove the view at certain index, which is not the same in Java since the order of items changed.

Test Plan:
Related issue and the simple empty react-native 0.37 app where it was tested:
https://github.com/facebook/react-native/issues/8968

In `manageChildren` in `NativeViewHierarchyManager.java` there was a loop by `indicesToRemove`, call `viewManager.removeViewAt(viewToManage, indexToRemove)`. However if the view is animated and it's ID in the array of `tagsToDelete`, do nothing and delegate this job of removing item to the loop by `tagsToDelete`.

Now we propagate `tagsToRemove` instead from `UIImplementation` to `NativeViewHierarchyOptimized` to `NativeViewHierarchyManager` and using it instead of `indicesToRemove`.

Also by having the old `indicesToRemove` loop we having the case that the *wrong* view is removed, but then, *correct* view is dropped in `tagsToDelete` loop. In result we have inconsistency when removing one view but dropping another. It's described in more details in mentioned above related issue comments.